Rename rustuna_storages crate to rustuna_storage
