+++ rustuna_pyo3/Cargo.toml
@@ -18,6 +18,6 @@ pyo3 = { version = "0.27", features = ["extension-module", "chrono"] }
 chrono = { workspace = true }
 rustuna_core = { workspace = true }
 rustuna_samplers = { workspace = true }
-rustuna_storages = { workspace = true }
+rustuna_storage = { workspace = true }
 rustuna_importance = { workspace = true }
 fanova = { workspace = true }
