+++ rustuna_storage/src/sqlite3_schema.sql
